fixing files & create text messages
