--- src/components/Chat/index.js
@@ -1,2 +1,0 @@
-export {default as ChatHeader} from "./Header";
-export {default as ChatFooter} from "./Footer";
